+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/TigerProxy.java
@@ -24,8 +24,10 @@ import de.gematik.test.tiger.proxy.tls.DynamicTigerKeyAndCertificateFactory;
 import de.gematik.test.tiger.proxy.tls.StaticTigerKeyAndCertificateFactory;
 import io.netty.handler.ssl.SslProvider;
 import jakarta.annotation.PreDestroy;
+import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.security.KeyManagementException;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
@@ -354,14 +356,7 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
 
     @Override
     public synchronized TigerRoute addRoute(final TigerRoute tigerRoute) {
-        tigerRouteMap.values().stream()
-            .filter(existingRoute ->
-                uriTwoIsBelowUriOne(existingRoute.getFrom(), tigerRoute.getFrom())
-                    || uriTwoIsBelowUriOne(tigerRoute.getFrom(), existingRoute.getFrom()))
-            .findAny()
-            .ifPresent(existingRoute -> {
-                throw new TigerProxyRouteConflictException(existingRoute);
-            });
+        assertThatRouteIsUnique(tigerRoute);
 
         log.info("Adding route {} -> {}", tigerRoute.getFrom(), tigerRoute.getTo());
         final Expectation[] expectations = buildRouteAndReturnExpectation(tigerRoute);
@@ -378,6 +373,17 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
 
         log.info("Created route {} with expectation {}", createdTigerRoute, expectations[0]);
         return createdTigerRoute;
+    }
+
+    private void assertThatRouteIsUnique(TigerRoute tigerRoute) {
+        tigerRouteMap.values().stream()
+            .filter(existingRoute ->
+                uriTwoIsBelowUriOne(existingRoute.getFrom(), tigerRoute.getFrom())
+                    || uriTwoIsBelowUriOne(tigerRoute.getFrom(), existingRoute.getFrom()))
+            .findAny()
+            .ifPresent(existingRoute -> {
+                throw new TigerProxyRouteConflictException(existingRoute);
+            });
     }
 
     private boolean uriTwoIsBelowUriOne(final String value1, final String value2) {
@@ -407,10 +413,19 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
     }
 
     private Expectation[] buildForwardProxyRoute(final TigerRoute tigerRoute) {
+        final URL url = buildUrlSafe(tigerRoute);
         return mockServerClient.when(request()
-                .withHeader("Host", tigerRoute.getFrom().split("://")[1])
-                .withSecure(tigerRoute.getFrom().startsWith("https://")))
+                .withHeader("Host", url.getAuthority())
+                .withSecure(url.getProtocol().equals("https")))
             .forward(new ForwardProxyCallback(this, tigerRoute));
+    }
+
+    private static URL buildUrlSafe(TigerRoute tigerRoute) {
+        try {
+            return new URL(tigerRoute.getFrom());
+        } catch (MalformedURLException e) {
+            throw new TigerProxyStartupException("Error while building route", e);
+        }
     }
 
     public void addAlternativeName(final String host) {